+++ sdk/python/tests/test_plugin.py
@@ -24,7 +24,7 @@ PLUGIN = """
 
 ---
 
-@type_definition
+@type
   id: type-epic
   type_name: epic
   description: "A large body of work"